Create a dtm converter
Sometimes it's usefull to draw a pseudo digital terrain model in sRGB.
But then, to use it as a test dataset for real applications you need to
have a floating point image with heightscale set correctly (1m is 1m).

One may have the same problem for heightmap, but it's rather unusual to
need a heightmap with a proper scale.